feat: add Hygen scaffolding for new endpoints
This commit introduces Hygen as a code generator for scaffolding boilerplate code for new endpoints.

To use this feature, run the command:
"yarn new-endpoint YourEndPointName"

This command will generate boilerplate files for the controller and service, and also inject the Controller and Service into api.module.ts.

This enhancement will significantly reduce the time and effort required to set up new endpoints in our application.